Created structor3 - a new shiny structor class
The new structor class is not perfect in that it is rather inefficient. This is because I lazily did nor bother to implement things efficiently, due to time management concerns. Some of the inefficiencies are probably easy to resolve, but the current methods are at least very robust, and maintains perceived syntax as normal to the user.
What's new?
- the new structor (v3) only stores data in the struct, and thus it is directly accessible to the user, maintaining all syntax regarding structs in MATLAB.
- the vector now retrieves all data from the struct every time it is queried. according to the current structure (mix / node-order).
- the structure is built then stored in a "map" variable, and is only recomputed when necessary. (whenever "str" changes in any way)
- Some special syntax for copying substructs is created.
This will probably be adopted as the main structor class soon.